refactor(h2r): rename HtmlToReactConverter to H2R
