+++ src/shared/components/H2R/parse/htmlMap.js
