+++ src/shared/components/H2R/parse/svgMap.js
